requirements.txt
# Dependencies for development and production
numpy
pandas
requests
python-dotenv
